Updated hsi to decouple the data, also restored the orignal sixpack so it can easily be compated to the 'screenbuilder'
